Django/Django_3/pyproject.toml
[tool.poetry]
name = "django_project"
version = "0.1.0"
description = "python django project"
authors = ["sub-blind <rlawotjq206@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
